+++ test/duckdb_test/appender_test.rb
@@ -6,22 +6,17 @@ require 'time'
 module DuckDBTest
   class AppenderTest < Minitest::Test
     def setup
-      @db = DuckDB::Database.open # FIXME
+      @db = DuckDB::Database.open
       @con = @db.connect
     end
 
     def safe_drop_table
-      @con.execute("DROP TABLE #{table};")
-    rescue DuckDB::Error
-      # ignore DuckDB::Error
+      @con.execute("DROP TABLE IF EXISTS #{table};")
     end
 
     def teardown
-      safe_drop_table
       @con.close
       @db.close
-    rescue DuckDB::Error
-      # ignore DuckDB::Error
     end
 
     def table
@@ -1323,6 +1318,48 @@ module DuckDBTest
       @con.query('CREATE SCHEMA s_override; CREATE TABLE s_override.t_ovr (id INT)')
       appender = nil
       _, err = capture_io { appender = @con.appender('wrong.t_ovr', schema: 's_override') }
+
+      assert_instance_of(DuckDB::Appender, appender)
+      assert_includes(err, 'deprecated')
+    end
+
+    def test_s_new_with_dot_notation_parses_schema_and_table
+      @con.query('CREATE SCHEMA s_dot; CREATE TABLE s_dot.t_dot (id INT)')
+      appender = nil
+      _, err = capture_io { appender = DuckDB::Appender.new(@con, 's_dot.t_dot') }
+
+      assert_instance_of(DuckDB::Appender, appender)
+      assert_includes(err, 'deprecated')
+    end
+
+    def test_s_new_with_3segment_dot_notation_parses_catalog_schema_table
+      @con.query("ATTACH ':memory:' AS ext_new_dot")
+      @con.query('CREATE TABLE ext_new_dot.main.t_new_dot (id INT)')
+      appender = nil
+      _, err = capture_io { appender = DuckDB::Appender.new(@con, 'ext_new_dot.main.t_new_dot') }
+
+      assert_instance_of(DuckDB::Appender, appender)
+      assert_includes(err, 'deprecated')
+    end
+
+    def test_s_new_with_double_quoted_table_name_is_treated_as_literal
+      @con.execute('CREATE TABLE "a.b" (col1 INTEGER)')
+      appender = DuckDB::Appender.new(@con, '"a.b"')
+
+      assert_instance_of(DuckDB::Appender, appender)
+    end
+
+    def test_s_new_with_single_quoted_table_name_is_treated_as_literal
+      @con.execute("CREATE TABLE 'a.b' (col1 INTEGER)")
+      appender = DuckDB::Appender.new(@con, "'a.b'")
+
+      assert_instance_of(DuckDB::Appender, appender)
+    end
+
+    def test_s_new_dot_notation_schema_keyword_overrides_dot_schema
+      @con.query('CREATE SCHEMA s_new_ovr; CREATE TABLE s_new_ovr.t_new_ovr (id INT)')
+      appender = nil
+      _, err = capture_io { appender = DuckDB::Appender.new(@con, 'wrong.t_new_ovr', schema: 's_new_ovr') }
 
       assert_instance_of(DuckDB::Appender, appender)
       assert_includes(err, 'deprecated')